アプリケーションファクトリの変更
