+++ Pipfile
@@ -14,6 +14,6 @@ flask = "*"
 python_version = "3.7"
 
 [scripts]
-start = "python app.py runserver"
+start = "python run.py runserver"
 format = "autopep8 -ivr ."
 lint = "flake8 --show-source --exclude .venv ."
