--- app.py
@@ -1,12 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-
-@app.route('/')
-def index():
-    return 'Hello, World'
-
-
-if __name__ == "__main__":
-    app.run()
