+++ application/__init__.py
@@ -1,0 +1,11 @@
+from flask import Flask
+
+
+def create_app():
+    app = Flask(__name__)
+
+    @app.route('/')
+    def index():
+        return 'Hello, World'
+
+    return app
